requirements.txt
.
fastapi
uvicorn[standard]
sse-starlette
feedparser
scipy
pyyaml
